Fix compilation in MSVC 2015.
* updated irrKlang to 1.4.0b
* fixed possible buffer overflow in debug code
* disabled SSE in Bullet (standard std::vector doesn't align memory properly)
* improved compatibility with Premake 5 (Premake 4 doesn't support MSVC 2015)

+++ shared/CTrigger.cpp
@@ -21,7 +21,7 @@ CTrigger::CTrigger(CLevel *pLevel, CInputBinaryStream &Stream):
     CObject(OFET_TRIGGER, pLevel), m_nWait(0)
 {
     m_iColGroup = COL_TRIGGER;
-    m_iColMask = COL_NONE;
+    m_iColMask = (short)COL_NONE;
     
     m_pColObj = new btCollisionObject();
     m_pColObj->setUserPointer(this);
